Update to match haf schema changes between 1.28.6-rc6 and -rc8

+++ db/main_loop.sql
@@ -362,7 +362,7 @@ STABLE
 AS
 $$
 DECLARE
-  __head_of_irreversible_block INT;
+  __head_of_irreversible_block BIGINT;
 BEGIN
